+++ packages/.monorepo/builder.ts
@@ -1,48 +1,12 @@
 import {build} from 'bun';
-import {existsSync} from 'node:fs';
 import path from 'path';
+import {validateExports} from './validate-exports';
 
 if (process.env.NODE_ENV !== 'production') {
 	throw new Error('This script must be run using NODE_ENV=production');
 }
 
 type Format = 'esm' | 'cjs';
-
-const validateExports = (
-	exports: Record<string, './package.json' | Record<string, string>>,
-) => {
-	const keys = Object.keys(exports);
-	for (const key of keys) {
-		const value = exports[key];
-		if (key === './package.json' && value === './package.json') {
-			continue;
-		}
-
-		if (typeof value === 'string') {
-			throw new Error(`Invalid export for ${key}`);
-		}
-
-		if (!value.import || !value.require || !value.module || !value) {
-			throw new Error(`Missing import or require for ${key}`);
-		}
-		const paths = Object.keys(value);
-		for (const entry of paths) {
-			if (
-				entry !== 'import' &&
-				entry !== 'require' &&
-				entry !== 'module' &&
-				entry !== 'types'
-			) {
-				throw new Error(`Invalid export: ${entry}: ${JSON.stringify(exports)}`);
-			}
-			const pathToCheck = path.join(process.cwd(), value[entry]);
-			const exists = existsSync(pathToCheck);
-			if (!exists) {
-				throw new Error(`Path does not exist: ${pathToCheck}`);
-			}
-		}
-	}
-};
 
 const getExternal = (deps: string[] | 'dependencies'): string[] => {
 	if (deps === 'dependencies') {
